Moved all plotting funcs to plottools file
